Moving shop_data out of merc.h
