+++ code/merc.h
@@ -378,22 +378,7 @@ struct help_data
 };
 
 
-//
-// Shop types.
-//
-
-#define MAX_TRADE					5
-
-struct shop_data
-{
-	bool fIsPetShop;
-	int pet_cage_vnum;
-	SHOP_DATA *next;			// Next shop in list
-	MOB_INDEX_DATA *pMobIndex;
-	short open_hour;			// First opening hour
-	short close_hour;			// First closing hour
-	short direction;			// exit dir
-};
+#include "entity/shop_data.h"
